docs: optimize demo docs
