+++ packages/docs/fluent-editor/demos/get-html.vue
@@ -1,10 +1,19 @@
 <script setup lang="ts">
 import { onMounted, ref } from 'vue'
-import katex from 'katex'
-import 'katex/dist/katex.min.css'
+
+// 代码块高亮
 import hljs from 'highlight.js'
 import 'highlight.js/styles/atom-one-dark.css'
+window.hljs = hljs
+
+// 插入公式
+import katex from 'katex'
+import 'katex/dist/katex.min.css'
 window.katex = katex
+
+// 截屏
+import Html2Canvas from 'html2canvas'
+window.Html2Canvas = Html2Canvas
 
 let editor
 
@@ -25,13 +34,16 @@ const mentionList = [
 
 const TOOLBAR_CONFIG = [
   ['undo', 'redo', 'clean', 'format-painter'],
-  [{ header: [1, 2, 3, 4, 5, 6, false] }, { size: ['12px', '14px', '16px', '18px', '20px', '24px', '32px', '36px', '48px', '72px'] }],
+  [{ header: [1, 2, 3, 4, 5, 6, false] }, { font: [] }, { size: ['12px', '14px', '16px', '18px', '20px', '24px', '32px', '36px', '48px', '72px'] }],
   ['bold', 'italic', 'strike', 'underline'],
   [{ color: [] }, { background: [] }],
   [{ align: [] }, { list: 'ordered' }, { list: 'bullet' }, { list: 'check' }],
+  [{ script: 'sub' }, { script: 'super' }],
+  [{ indent: '-1' }, { indent: '+1' }],
+  [{ direction: 'rtl' }],
   ['link', 'blockquote', 'code', 'code-block'],
   ['image', 'file', 'better-table'],
-  ['emoji', 'video', 'formula'],
+  ['emoji', 'video', 'formula', 'screenshot'],
 ]
 
 const articleRef = ref<HTMLElement>()
