+++ packages/docs/fluent-editor/docs/image-upload.md
@@ -1,11 +1,15 @@
 # 图片上传
 
+## 基本用法
+
 通过配置工具栏按钮 `image`，可以开启图片上传功能，图片支持等比放大/缩小。
 
 :::demo src=demos/image-upload.vue
 :::
 
 ## 上传到服务器
+
+配置 `uploadOption.imageUpload` 可以将图片上传到服务器，并将服务器返回的图片 URL 渲染到编辑器中。
 
 :::demo src=demos/image-upload-to-server.vue
 :::
